requirements.txt
playwright==1.58.0
python-dotenv==1.0.0
pandas==2.3.3
openpyxl==3.1.5
lxml==6.0.2
